backend/requirements.txt
fastapi
uvicorn
gunicorn
aiosqlite
asyncpg
sqlalchemy
alembic
pydantic
pydantic-settings
python-jose[cryptography]
passlib[bcrypt]
python-multipart
slowapi
redis
python-dotenv
httpx
pytest
pytest-asyncio
sentry-sdk
